dev-back/pyproject.toml
[project]
name = "dev-back"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.128.0",
    "uvicorn>=0.32.0",
    "httpx>=0.27.0",
    "python-jose[cryptography]>=3.3.0",
    "python-multipart>=0.0.12",
    "pydantic-settings>=2.6.0",
    "sqlalchemy>=2.0.0",
    "alembic>=1.13.0",
    "asyncpg>=0.29.0",
    "python-dotenv>=1.0.0",
    "psycopg2-binary>=2.9.11",
]
